requirements.txt
APScheduler>=3.10.4
flask
PyYAML>=6.0.1
SQLAlchemy>=2.0.30
playwright>=1.44.0
playwright-stealth>=1.0.6
